+++ src/shaders/scene.glsl
@@ -7,6 +7,7 @@ layout(set = 0, binding = 0, std140) uniform sdf_view_Scene {
     vec4 light_direction;
     vec4 light_color;
     vec4 background;
+    vec4 object_color;
 } sdf_view_scene;
 #define sdf_view_resolution sdf_view_scene.resolution.xy
 #define sdf_view_sample_grid int(sdf_view_scene.resolution.z)